create Membrane.Core.Telemetry for telemetry functions
